pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.37.23",
    "cryptography>=44.0.2",
    "django>=5.1.7",
    "django-allauth>=65.6.0",
    "django-storages>=1.14.5",
    "pillow>=11.1.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
